requirements.txt
streamlit
langchain
langchain-openai
langchain-groq
langchain-google-genai
langchain-core
PyPDF2
faiss-cpu
unstructured
python-dotenv
tiktoken
requests
google-api-core
googleapis-common-protos
protobuf
google-generativeai
langchain-community
sentence-transformers
python-docx
pandas


